backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
python-multipart
pyjwt
pytest
httpx
exceptiongroup
bcrypt
itsdangerous
pydantic[email]